Implemented offset functionality
You can now set a offset in the parameters of the distance in the file you are reading from, the output file size will be the diference on the bytes of the input file size - the offset. To be implemented in the futuro a offset counting from the end too, so beware :)
This is in experimental so if you find bugs dont be surprised, and please contact me about them.