Implementing threads
DO NOT TEST THIS. IF YOU DO IM NOT RESPONSIBLE.
Ok, now that i scared off the normies, here is what's happening.
The code "works" but it's still very (and i mean very) experimental. Im still activily working on it when i have time (so like 3 times a week, 4), so ill soon upload a more stable version of this code. THE CODE STILL HAS BUGS, so expect UB at times, I only upload the code i make that works for a specific test i made (i use this to store the current version of the code, and when it's finished, it will store the final version of it (: !) so keep that in mind. Anny questions regarding the code you can ask me in discord, tag in bio.